links
was trying to make it so that the links take you to the green part of
the page but its going to be more complicated than i thought because
php means that you have to refresh the page to update the content in
the green bit. but I think I know how we can fix this. we need to pass
a get variable through the links so that when the page reloads the
margin is set to -80%. will not be possible through jquery as we are
not using AJAX :)