activation
so like i mentioned before you cant contact the database unless your in
it so this php file only works when its on janis is server. so if you
go to janis0809.keaweb.dk/random-exp and use a valid code eg. abc00
then this works. there are no images because i didnt upload the images
folder because that would have used too much of my phone’s data :)  i
think thats it